Implement Interval Tree (#204)
Uses an interval tree to query which line in the file an offset range corresponds to. Previously this was determined by looping over an array. This is heavily used to determine the cursor position during the prettify phase.

This isn't really an issue when a file is a couple thousand lines, but when formatting a large repo this should provide a substantial speedup since it's a O(n) -> O(lgn) improvement.

The string formatted is the contents of `runtests.jl` (~4600 lines)

**old**

```
BenchmarkTools.Trial:
  memory estimate:  13.70 MiB
  allocs estimate:  223063
  --------------
  minimum time:     26.294 ms (0.00% GC)
  median time:      27.123 ms (0.00% GC)
  mean time:        28.359 ms (4.93% GC)
  maximum time:     34.134 ms (17.61% GC)
  --------------
  samples:          177
  evals/sample:     1

```

**new**

```
BenchmarkTools.Trial:
  memory estimate:  15.20 MiB
  allocs estimate:  223103
  --------------
  minimum time:     15.131 ms (0.00% GC)
  median time:      15.704 ms (0.00% GC)
  mean time:        17.318 ms (9.88% GC)
  maximum time:     24.571 ms (30.20% GC)
  --------------
  samples:          289
  evals/sample:     1
```