fix #202
In instances with multiple consecutive `=`, sometimes unnesting
would remove the newline in an inner `=` call but the indentation
added during the initial nest should stay. This caused extra spacing,
i.e. `a =     b` instead of `a = b`.

Now during unnesting, if the node is a BinaryOpCall, the additional
indent is always reset.