Resolve #262 (#265)
* Resolve #262

During the pass to move @ to the end of the macroname, the FST walk was
descending into inner macro calls. This is problematic since all leaf
nodes with the value "@" were removed. This is solved by modifying `walk` to stop descending into a subtree if value other than `nothing` is returned from the passed function.

* fmt

* walk doc + remove old code

* fix test

* handle case when @ is in the middle of a call chain

* fmt