Correctly increment state offset when unicode is encountered
Alignment requires knowing the line offset for characters in the original file.
The way Tokenize.jl and, in turn CSTParser.jl handle unicode is not compatible
with the formatters notion of length/margin.

The key issue being a unicode character can span multiple bytes yet only
take up a single characters space in a file. The formatter cares about
the latter case while the aforementioned packages care about the former.

To overcome this the the state offset is refactored to increment by number
of characters instead of number of bytes.