YASStyle vcat/typedvcat fix
This used to fallback to DefaultStyle. Now it aligns the argument to one
whitespace after the `[` and args are only joined to the same line only if
that was the case in the original source. If the arguments on the same
line exceed the maximum margin the arguments are split unto different
lines.