YASStyle vcat/typedvcat fix (#357)
* YASStyle vcat/typedvcat fix

This used to fallback to DefaultStyle. Now it aligns the argument to one
whitespace after the `[` and args are only joined to the same line only if
that was the case in the original source. If the arguments on the same
line exceed the maximum margin the arguments are split unto different
lines.

* Placeholders are not always put in for vcat and typedvcat so we need to
check if the node is a Newline.

* additional check

* uncomment non-YAS test files

* fmt

* rm old code

* v0.13.0

* kw syntax