Resolve #352
For BlueStyle fix it was literally just a typo of 0 instead of -1 in the
dispatch for non-default styles

For YASStyle we have to copy-pasta the nest! function for Vector{EXPR}
into n_for!(::YASStyle, ...) so that we can set the indent for the
argument block that sometimes occurs in a for/while loop.

In generator expressions, the for loop is not parsed as as a for loop
but rather a vector of expressions that happens to have a `for` keyword
which is why this worked fine for those.