fix #387
Handles the edge case where the first argument of a Call node
is nestable. For example:

```
new{T1,T2}(arg1,arg2)
```

In this case if the indent is set to the offset of the first argument after
`(` and `new{T1,T2}` is nested, the indentation will be incorrect.

before (aligned to after the initial position of `(`):

```
new{T1,
    T2}(arg1,
           arg2)
```

after (aligned to after new position of `(`):

```
new{T1,
    T2}(arg1,
        arg2)
```