Fix #417
Macro strings: handle the case where the identifier is part of a
quotenode(s), i.e. `A.foo"a"` along with `foo"a"`.

Since the identifier is a special case we add a prettification function
for it `p_macrostr_identifier` which is only called very specific
circumstances, meaning it's not included in the main `pretty` function.