fix #512
Trims lines in a multiline comment that are whitespace with a newline
at the end to simply that newline. Otherwise, the amount of whitespace prior to the
start of the multiline comment is prepended to that line causing the
whitespace before the newline to increase on each format.

For example (whitespace in mutliline comment is noted by <ws>)

begin
    #=
    <ws><ws>\n
    =#
end

Since there are 4 spaces before #= on format 4 more spaces will be added
to the 2nd line in the mutliline comment, meaning "<ws><ws>\n" becomes
"<ws><ws><ws><ws><ws><ws>\n" and so on.