fix #546
1. The startline/endline for the created "return" keyword was using the
range of the block node instead of the last node in the block. Odd this
didn't pop up as a bug sooner but I guess since the following node was
automatically joined to the return node it hasn't been problematic.
Nonetheless `join_lines_based_on_source` brings this issue to light so
we need to deal with it.
2. `walk` mutates the `line_offset` of the `State` and we weren't resetting
it properly in `prepend_return!` which could cause odd indentation,
seemingly only when `join_lines_based_on_source` is `true`. It's now
properly reset so this is no longer an issue.