fix #543
This problem occurs when a line in the original source code contains >=
2 rows.

The current format breaks up the rows into individual lines but it
prepends an undesirable amount of whitespace, based on the length of the
previous rows on that line.

The fix is to just not prepend whitespace if the row is not the initial
row on that line.

I'm not sure this is the "best" fix but it's certainly better than what
the current format by some margin (pun intended).