Multiline comment corner case + better error message on formatted text (#572)
error

Unfortunately it looks like the the position the error occurred is hard
to derive from CSTParser's ParseState so for we'll at least number the
formatted output. Although without accurate line info I'm not sure how
helpful that actually is.

In slightly better news we've added a temporary measure to handle a
corner case of multiline comments being moved to the end of the Lines
which can apparently cause havoc for CSTParser. Luckily this canbe
alleviated by prepending a single space prior to the comment. It's
pretty hacky but better than erroring for no good reason!