`format_docstrings` handle REPL usage in all cases
Before we assumed if the REPL was used in a docstring
it would be surrounded by a block such as:

```jldoctest
...julia code...
```

but it would only work if the type was `jldoctest`.

This PR assumes all julia-esque code blocks can contain REPL usage and
handles it accordingly.

As a bonus it looks like the function is a bit simpler as well.

fix #597