Make `format_file` an alias for `format`
At the moment when `format_file` is called it does not look for options set
in a .JuliaFormatter.toml as `format` does. This can cause different
outputs based on whether you call either `format` or `format_file` which
should not be the case.

This PR renames the old `format_file` to `_format_file` and
then has `format_file` be an alias to `format` so that it works as
intended.