Fix `align_assignment` to work with nested macros
For example:

```julia
@foo @noinline Base.@constprop :none aaaaaaaaaaaaa() = 0
@foo @noinline Base.@constprop :none bbbbbbbbbbbbbbbbbbb()     = 0
@foo @foo @ccccccccccccccccccccccccccccccccccccccccccccccccc() = 0
```

Previously this would throw an error (which was uncaught) since the binary op call was not
found at the top level.

With this PR we search until we reach a binary op call or there are no
more nodes and then align to that binary op. Additionally we store the
reference to the exact node which will be aligned in the AlignGroup.
This small change saves time in finding the node again later on.